--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/FlyBehavior.java
@@ -1,7 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public interface FlyBehavior {
-
-    void fly();
-
-}
