--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/FlyNoWay.java
@@ -1,10 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public class FlyNoWay implements FlyBehavior {
-
-    @Override
-    public void fly() {
-        System.out.println("I can't fly");
-    }
-
-}
