--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/FlyWithWings.java
@@ -1,10 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public class FlyWithWings implements FlyBehavior {
-
-    @Override
-    public void fly() {
-        System.out.println("I'm flying!!");
-    }
-
-}
