+++ headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/Main.java
@@ -1,8 +1,11 @@
 package com.headfirst.DesignsPatterns.ChapterOne;
 
+import com.headfirst.DesignsPatterns.ChapterOne.classes.MallardDuck;
+import com.headfirst.DesignsPatterns.ChapterOne.interfaces.IDuck;
+
 public class Main {
     public static void main(String[] args) {
-        Duck mallard = new MallardDuck();
+        IDuck mallard = new MallardDuck();
         mallard.performQuack(); // Quack
         mallard.performFly(); // I'm flying!!
     }
