--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/MuteQuack.java
@@ -1,10 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public class MuteQuack implements QuackBehavior {
-
-    @Override
-    public void quack() {
-        System.out.println("<< Silence >>");
-    }
-
-}
