--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/Quack.java
@@ -1,10 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public class Quack implements QuackBehavior {
-
-    @Override
-    public void quack() {
-        System.out.println("Quack");
-    }
-
-}
