--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/QuackBehavior.java
@@ -1,7 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public interface QuackBehavior {
-
-    void quack();
-
-}
