--- headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/Squeak.java
@@ -1,10 +1,0 @@
-package com.headfirst.DesignsPatterns.ChapterOne;
-
-public class Squeak implements QuackBehavior {
-
-    @Override
-    public void quack() {
-        System.out.println("Squeak");
-    }
-
-}
