+++ headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/classes/Squeak.java
@@ -1,0 +1,12 @@
+package com.headfirst.DesignsPatterns.ChapterOne.classes;
+
+import com.headfirst.DesignsPatterns.ChapterOne.interfaces.IQuackBehavior;
+
+public class Squeak implements IQuackBehavior {
+
+    @Override
+    public void quack() {
+        System.out.println("Squeak");
+    }
+
+}
