+++ headfirst/src/main/java/com/headfirst/DesignsPatterns/ChapterOne/interfaces/IQuackBehavior.java
@@ -1,0 +1,7 @@
+package com.headfirst.DesignsPatterns.ChapterOne.interfaces;
+
+public interface IQuackBehavior {
+
+    void quack();
+
+}
